backend/requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
google-genai
boto3
python-multipart
fal-client
moviepy
imageio-ffmpeg
httpx
pillow
moviepy
timezonefinder
PyJWT
google-auth
duckduckgo-search
beautifulsoup4
lxml
